+++ src/components/SCheckbox/SCheckbox.vue
@@ -12,6 +12,7 @@ const emit = defineEmits<{
 const props = defineProps<{
   disabled?: boolean;
   id?: string;
+  inline?: boolean;
   modelValue: boolean;
   name?: string;
   reverse?: boolean;
@@ -33,12 +34,15 @@ const computedId = computed(() => props.id ?? uuidv4());
 </script>
 
 <template>
-  <div :class="['flex w-full', (!hasSlotContent($slots.default) || !hasSlotContent($slots.description)) && 'items-center', reverse ? 'flex-row-reverse justify-between' : ' gap-3']">
+  <div
+    :class="[
+      'flex w-full gap-3',
+      reverse && 'flex-row-reverse justify-between',
+      disabled && 'opacity-50 pointer-events-none',
+    ]"
+  >
     <input
-      :class="[
-        'border border-gray-300 bg-white focus:s-ring focus:ring-offset-0 rounded text-primary-600 accent-primary-600 cursor-pointer',
-        disabled && 'opacity-50 pointer-events-none',
-      ]"
+      class="border border-gray-300 bg-white focus:s-ring focus:ring-offset-0 rounded text-primary-600 accent-primary-600 cursor-pointer"
       type="checkbox"
       :disabled="disabled"
       :id="computedId"
@@ -49,17 +53,27 @@ const computedId = computed(() => props.id ?? uuidv4());
     />
     <div
       v-if="hasSlotContent($slots.default) || hasSlotContent($slots.description)"
-      :class="['flex flex-col', $slots.default && $slots.description && 'gap-1']"
+      :class="[
+        'flex flex-col justify-center',
+        $slots.default && $slots.description && 'gap-1',
+        hasSlotContent($slots.description) ? '-mt-1' : '-mt-0.5',
+      ]"
     >
-      <label
-        v-if="hasSlotContent($slots.default)"
-        :for="computedId"
-        class="text-gray-900 font-semibold text-sm leading-3"
-        ><slot
-      /></label>
-      <p v-if="hasSlotContent($slots.description)" class="text-sm text-gray-500 font-normal">
+      <label v-if="hasSlotContent($slots.default) && !inline" :for="computedId" class="label">
+        <slot />
+      </label>
+      <p v-if="hasSlotContent($slots.description)" :class="['text-sm text-gray-500 font-normal', inline && '']">
+        <label v-if="hasSlotContent($slots.default) && inline" :for="computedId" class="label">
+          <slot />
+        </label>
         <slot name="description" />
       </p>
     </div>
   </div>
 </template>
+
+<style>
+.label {
+  @apply text-gray-900 font-semibold text-sm;
+}
+</style>